refactor: move `FractalView` to a dedicated module
