+++ src/fractal/mod.rs
@@ -3,5 +3,4 @@ pub mod iter;
 pub mod render;
 pub mod types;
 
-pub use render::*;
 pub use types::{Fractal, Set};
